Refixed the data protection key issue, this time using Azurite on my local as opposed to the file stuff
